+++ app/components/annotations/annotations-directives.js
@@ -6,14 +6,14 @@
 angular.module('annotations')
   .directive('annotations', [function () {
     var link = function (scope, element, attrs) {
-      scope.annotations = {};
+      scope.annotations = [];
     };
 
     return {
       restrict: 'E',
       scope: {
-        asset: '=',
-        timeState: '='
+        data: '=',
+        timeState: '=',
       },
       link: link,
       templateUrl: 'annotations/templates/annotations.html'
@@ -49,24 +49,40 @@ angular.module('annotations')
       };
 
       /**
-       * Get all annotations for an asset.
+       * Get all annotations for an asset or geometry.
        * @returns {array} - An array of annotations.
        */
       var fetchAnnotations = function() {
-        AnnotationsService.getAnnotationsForObject(
-          scope.asset.entity_name,
-          scope.asset.id,
-          5,
-          scope.timeState.start,
-          scope.timeState.end,
-          fetchAnnotationsSuccess
-        );
+        if (scope.data.properties && scope.data.properties.annotations) {
+          var events = scope.data.properties.annotations.data;
+          var annotations = [];
+          events.forEach(function (ev) {
+            annotations.push(ev.properties);
+          });
+          fetchAnnotationsSuccess(annotations);
+        }
+
+        else if (scope.data && scope.data.entity_name && scope.data.id) {
+          AnnotationsService.getAnnotationsForObject(
+            scope.data.entity_name,
+            scope.data.id,
+            5,
+            scope.timeState.start,
+            scope.timeState.end,
+            fetchAnnotationsSuccess
+          );
+        }
+
+        else {
+          scope.annotations = [];
+        }
+
       };
 
       /**
        * Get annotations when asset changes.
        */
-      scope.$watch('asset', function () {
+      scope.$watch('data', function () {
         fetchAnnotations();
       });
 
@@ -80,6 +96,7 @@ angular.module('annotations')
       });
       /**
        * Update the front-end to reflect a successful delete of an annotation.
+       * Both on the map and timeline as in the box.
        * @param {object} id - The ID of the asset.
        * @param {?} value - Not actually used but required by $resource.
        * @param {dict} responseHeaders - Not actually used but required
@@ -88,6 +105,7 @@ angular.module('annotations')
       var deleteAnnotationSuccess = function(
           annotation, value, responseHeaders) {
         scope.annotations.splice(scope.annotations.indexOf(annotation), 1);
+        AnnotationsService.refreshAnnotationLayer();
       };
 
       /**
@@ -127,7 +145,7 @@ angular.module('annotations')
       link: link,
       restrict: 'E',
       scope: {
-        asset: '=',
+        data: '=',
         annotations: '=',
         timeState: '='
       },
@@ -235,6 +253,7 @@ angular.module('annotations')
        */
       var createAnnotationSuccess = function(value, responseHeaders){
         scope.annotations.splice(0, 0, value);
+        AnnotationsService.refreshAnnotationLayer();
       };
 
       /**
@@ -258,7 +277,7 @@ angular.module('annotations')
        */
       scope.createAnnotation = function () {
         AnnotationsService.addAnnotationToObject(
-          scope.asset,
+          scope.data,
           scope.text,
           scope.attachment,
           scope.timelineat,
@@ -272,7 +291,7 @@ angular.module('annotations')
       link: link,
       restrict: 'E',
       scope: {
-        asset: '=',
+        data: '=',
         annotations: '=',
         timelineat: '='
       },