+++ app/components/user-menu/user-menu-directive.js
@@ -4,8 +4,10 @@
  */
 angular.module('user-menu')
   .directive('userMenu', function () {
-   
-  var link = function () {};
+
+  var link = function () {
+    Lizard.startPlugins(); // jshint ignore:line
+  };
 
   return {
     restrict: 'E',