+++ CHANGES.rst
@@ -44,6 +44,8 @@ annotation
 
 - Fix bugs in zoom buttons, openLayerGroups and clearing of the search query.
 
+- Initial restyle of the data menu.
+
 
 Release 2.6.2 (2016-1-18)
 ---------------------